New format based on dicts implemented server-side
This doesn't break the compatibility with existing script versions.
However, it completely changes how the server is implemented and changes
the Web page a little bit.

Scores storage was previously implemented as a JSON file containing a
list of 3-elements lists: user, score and text. We now have a
`ScoresStorage` object which contains a bunch of dict-like `Score`
objects. This allows us to change the underlying storage without
breaking `ScoresStorage`’s interface, and to add other attributes on
scores like the upload date (which is not implemented right now).

This new version will automatically update any old storage file with the
new format, there’s no need to use `utils/updatefmt.py`, which will be
removed in a later commit.

While this commit marks a “working” version of the new storage, not all
tests have been adapted, and thus I won’t release the 0.0.7 version
right now.